Partial Refactor
Partial refactor of Visualizer, Calculator, AdvancedCalculator, and
Render3D to eliminate dependency on Visualizer.
Refactored Calculator to act as an actual graphing calculator to
simplify how it is utilized and to create some encapsulation. Modified
Visualizer to work with the changes.
Changed names of Calculator and AdvancedCalculator to GraphingCalculator
and AdvancedGraphingCalculator respectively.
Fixed system test bug #1